@@ -1,29 +1,110 @@
-import React, { useEffect } from 'react';
-import { TextContainer, EntriesContainer } from '../styled-components';
+import React, { useEffect, useMemo, useState } from 'react';
+import styled from "styled-components";
 import EntryCard from './EntryCard';
 import { useSelector, useDispatch } from 'react-redux';
 import { getEntries } from '../actions/entry-actions';
+import SleepChart from './SleepChart';
+import moment from "moment";
+import SleepScoreCard from './SleepScoreCard';
+import SleepScheduleBar from './SleepScheduleBar';
 
 const Dashboard = () => {
-    const {entries, isFetching, error} = useSelector(state => state.entriesReducer);
+    const { entries } = useSelector(state => state.entriesReducer);
     const dispatch = useDispatch();
 
-    useEffect(() => {
-        dispatch(getEntries());
-    }, [dispatch]);
+    
+
+    const getLast7Days = () => {
+        const today = moment().endOf("day");
+        const sevenDaysAgo = moment().startOf("day").subtract(6, "days");
+
+        const filtered = entries.filter(entry => moment(entry.sleep_start).isBetween(sevenDaysAgo, today, null, []));
+        return filtered;
+    };
+
+    const last7Days = useMemo(() => getLast7Days(), [entries]);
+    
+    const getLatestEntry = () => {
+        if (!entries || entries.length === 0) return null;
+
+        return entries.reduce((latest, entry) => moment(entry.sleep_start).isAfter(moment(latest.sleep_start)) ? entry: latest
+        );
+    }
+    const latestEntry = useMemo(() => getLatestEntry(), [entries]);
+        
     
     return (
-        <>
-            <TextContainer>
-                <h3>Hours Slept</h3>
-            </TextContainer>
+        <DashboardGrid>
+          <Row>
+            <ScoreCardContainer>
+                <SleepScoreCard data = {last7Days} />
+            </ScoreCardContainer>
+                    
+            <DurationChartContainer>
+                <SleepChart data = {last7Days} />
+            </DurationChartContainer>
+          </Row>
+          <Row>
+            <ScheduleBarContainer>
+                <SleepScheduleBar data = {last7Days} />
+            </ScheduleBarContainer>
 
-            <EntriesContainer>
-                <h5>Sleep Entries</h5>
-                {entries.map(entry => <EntryCard key = {entry.id} {...entry} />)}
-            </EntriesContainer>
-        </>
+            <SleepEntryContainer>
+                {latestEntry && <EntryCard key = {latestEntry.id} {...latestEntry} />}
+            </SleepEntryContainer>
+          </Row> 
+            
+        </DashboardGrid>
     );
 };
 
 export default Dashboard;
+
+
+const DashboardGrid = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+  width: 100%;
+  max-width: 1100px;
+  margin: auto;
+`;
+
+const Row = styled.div`
+  display: flex;
+  gap: 1.5rem;
+  flex-wrap: wrap;
+`;
+
+const ScoreCardContainer = styled.div`
+  flex: 1;
+  max-width: 280px;
+  background: #fff;
+  border-radius: 1rem;
+  width: 100%;
+  height: 350px;
+`;
+
+const DurationChartContainer = styled.div`
+  flex: 2;
+  background: #fff;
+  border-radius: 1rem;
+  width: 100%;
+  height: 350px;
+`;
+
+const SleepEntryContainer = styled.div`
+  flex: 1;
+  background: #fff;
+  border-radius: 1rem;
+  width: 100%;
+  height: 100%;
+`;
+
+const ScheduleBarContainer = styled.div`
+  flex: 1;
+  background: #fff;
+  border-radius: 1rem;
+  width: 100%;
+  height: 100%;
+`;
