@@ -1,10 +1,11 @@
 import Axios from "axios";
+import { useEffect, useState } from "react";
 
 export default function AxiosWithAuth() {
     const token = localStorage.getItem("sleepToken");
 
     return Axios.create({
-        baseURL: "https://webpt15-sleep-tracker-api.herokuapp.com/api",
+        baseURL: "https://sleeptracker-back-end.onrender.com/api",
         headers: {
             Authorization: `Bearer ${token}`,
         }
